requirements.txt
Flask==3.0.0
requests
python-dotenv
notion-client==2.2.1
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-generativeai
rich==14.0.0
tqdm==4.67.1
pandas
openpyxl
xlsxwriter
Markdown